# apps/backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SaveMate API",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:4000"
    }
  ],
...
  "paths": {
...
    "/deals": {
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "required": false,
            "name": "page",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50,
              "default": 20
            },
            "required": false,
            "name": "limit",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "categoryId",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "city",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "DOLNOSLASKIE",
                "KUJAWSKO_POMORSKIE",
                "LUBELSKIE",
                "LUBUSKIE",
                "LODZKIE",
                "MALOPOLSKIE",
                "MAZOWIECKIE",
                "OPOLSKIE",
                "PODKARPACKIE",
                "PODLASKIE",
                "POMORSKIE",
                "SLASKIE",
                "SWIETOKRZYSKIE",
                "WARMINSKO_MAZURSKIE",
                "WIELKOPOLSKIE",
                "ZACHODNIOPOMORSKIE"
              ]
            },
            "required": false,
            "name": "voivodeship",
            "in": "query"
          },
          {
            "schema": {
              "type": "number",
              "nullable": true
            },
            "required": false,
            "name": "minPrice",
            "in": "query"
          },
          {
            "schema": {
              "type": "number",
              "nullable": true
            },
            "required": false,
            "name": "maxPrice",
            "in": "query"
          },
          {
            "schema": {
              "type": "number",
              "nullable": true,
              "minimum": 0,
              "maximum": 100
            },
            "required": false,
            "name": "discountMin",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "maxLength": 200
            },
            "required": false,
            "name": "q",
            "in": "query"
          },
          {
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              ]
            },
            "required": false,
            "name": "tags",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "required": false,
            "name": "dateFrom",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "required": false,
            "name": "dateTo",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "newest",
                "biggestDiscount",
                "lowestPrice"
              ],
              "default": "newest"
            },
            "required": false,
            "name": "sort",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "minLength": 1
                          },
                          "title": {
                            "type": "string",
                            "minLength": 1
                          },
                          "price": {
                            "type": "number",
                            "minimum": 0
                          },
                          "originalPrice": {
                            "type": "number",
                            "minimum": 0
                          },
                          "discountPercent": {
                            "type": "integer",
                            "minimum": 0,
                            "maximum": 100
                          },
                          "status": {
                            "type": "string",
                            "enum": [
                              "DRAFT",
                              "PENDING",
                              "APPROVED",
                              "REJECTED",
                              "EXPIRED"
                            ]
                          },
                          "city": {
                            "type": "string",
                            "minLength": 1
                          },
                          "voivodeship": {
                            "type": "string",
                            "enum": [
                              "DOLNOSLASKIE",
                              "KUJAWSKO_POMORSKIE",
                              "LUBELSKIE",
                              "LUBUSKIE",
                              "LODZKIE",
                              "MALOPOLSKIE",
                              "MAZOWIECKIE",
                              "OPOLSKIE",
                              "PODKARPACKIE",
                              "PODLASKIE",
                              "POMORSKIE",
                              "SLASKIE",
                              "SWIETOKRZYSKIE",
                              "WARMINSKO_MAZURSKIE",
                              "WIELKOPOLSKIE",
                              "ZACHODNIOPOMORSKIE"
                            ]
                          },
                          "categoryId": {
                            "type": "string",
                            "minLength": 1
                          },
                          "tags": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          },
                          "validFrom": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "validTo": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          }
                        },
                        "required": [
                          "id",
                          "title",
                          "price",
                          "originalPrice",
                          "discountPercent",
                          "status",
                          "city",
                          "voivodeship",
                          "categoryId",
                          "tags",
                          "validFrom",
                          "validTo",
                          "createdAt"
                        ]
                      }
                    },
                    "page": {
                      "type": "object",
                      "properties": {
                        "page": {
                          "type": "integer",
                          "minimum": 1
                        },
                        "limit": {
                          "type": "integer",
                          "minimum": 1
                        },
                        "total": {
                          "type": "integer",
                          "minimum": 0
                        },
                        "totalPages": {
                          "type": "integer",
                          "minimum": 0
                        }
                      },
                      "required": [
                        "page",
                        "limit",
                        "total",
                        "totalPages"
                      ]
                    }
                  },
                  "required": [
                    "items",
                    "page"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_ERROR",
                            "UNAUTHORIZED",
                            "FORBIDDEN",
                            "NOT_FOUND",
                            "CONFLICT",
                            "RATE_LIMITED",
                            "INTERNAL"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "nullable": true
                        },
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ]
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_ERROR",
                            "UNAUTHORIZED",
                            "FORBIDDEN",
                            "NOT_FOUND",
                            "CONFLICT",
                            "RATE_LIMITED",
                            "INTERNAL"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "nullable": true
                        },
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ]
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  }
}